app/requirements.txt
django==4.2.7
pyTelegramBotAPI==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot-pagination==0.0.3
python-dotenv==1.0.0